requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
autopep8==2.3.2
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-default-group==1.2.4
colorama==0.4.6
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.1
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
filetype==1.2.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.9.0
google-api-core==2.26.0
google-auth==2.41.1
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.0.2
langchain-core==1.0.0
langchain-google-genai==3.0.0
langfuse==3.8.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.37
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.9.0
numpy==2.3.4
openai==2.6.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.3.1
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.33.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pydantic==2.12.3
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rpds-py==0.27.1
rsa==4.9.1
sentry-sdk==2.45.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlite-fts4==1.0.3
sqlite-utils==3.38
starlette==0.48.0
streamlit==1.50.0
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
